backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.110.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
limits==5.8.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.41.5
pytest==7.4.4
pytest-asyncio==0.21.1
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.15.0
uvicorn==0.29.0
wrapt==2.1.1
websockets==12.0
celery==5.3.6
redis==5.0.1
fastapi-mail==1.4.1
fastapi-cache2-redis==0.2.1
asyncio-redis==0.16.0